+++ src/components/ErrorMessage/ErrorMessage.tsx
@@ -6,8 +6,6 @@ export interface ErrorMessageProps {
   /** Feilmelding */
   children: JSX.Element | string;
   showError?: boolean;
-  /** @deprecated Do not use */
-  ariaLabel?: string;
   /** Overstyring av stiler */
   className?: string;
 }
@@ -17,7 +15,7 @@ export interface ErrorMessageProps {
  */
 export const ErrorMessage: React.FC<ErrorMessageProps> = (props) => {
   const errorClassNames = getClassNames();
-  const { ariaLabel, className } = props;
+  const { className } = props;
   const showError = props.showError !== undefined ? props.showError : true;
   if (!showError) {
     return null;
@@ -26,7 +24,6 @@ export const ErrorMessage: React.FC<ErrorMessageProps> = (props) => {
     <span
       role="alert"
       aria-atomic="true"
-      aria-label={ariaLabel}
       className={classnames(errorClassNames, className)}
     >
       {props.children}